requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.1.0
colorama==0.4.6
deep-translator==1.11.4
diffusers==0.32.2
distro==1.9.0
duckduckgo_search==7.4.4
fastapi==0.115.8
google-pasta==0.2.0
groq==0.18.0
huggingface-hub==0.29.1
Jinja2==3.1.5
keras==3.8.0
Markdown==3.7
numpy==2.0.2
openai==1.63.2
pandas==2.2.3
pgvector==0.3.6
phidata==2.7.10
psycopg==3.2.4
psycopg-binary==3.2.4
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.8.0
SQLAlchemy==2.0.38
streamlit==1.42.2
sympy==1.13.1
tensorboard==2.18.0
tensorflow==2.18.0
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
uvicorn==0.34.0
